# openapi.yaml
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
info:
  title: FastAPI
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /drone_tracking_alert/{droneTrackingId}:
    get:
      operationId: get_drone_tracking_alert_droneTrackingId_drone_tracking_alert__droneTrackingId__get
      parameters:
      - in: path
        name: droneTrackingId
        required: true
        schema:
          title: Dronetrackingid
          type: string
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Drone Tracking Alert Dronetrackingid
      tags:
      - Extra Routes
      x-description: Given a Drone tracking object with basic telemetry data, it verifies
        that the coordinates of a given timestamp are correct, if not, it sets an
        alarm to the drone system.
  /drone_tracking_alert_from_drone/{droneId}:
    get:
      operationId: get_drone_tracking_alert_from_drone_droneId_drone_tracking_alert_from_drone__droneId__get
      parameters:
      - in: path
        name: droneId
        required: true
        schema:
          title: Droneid
          type: string
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Drone Tracking Alert From Drone Droneid
      tags:
      - Extra Routes
      x-description: Given a Drone id, check the its last tracking data available
        with basic telemetry data, it verifies that the coordinates of a given timestamp
        are correct, if not, it sets an alarm to the drone system.
...